Refactor: Remove oracle_node_app package and move classes to oracle_node
